APPLYING INTEL PATCHES sent to Julie on Feb 19th 2016 by Dima from INTEL (dmitry.g.baksheev@intel.com)
[PATCH 15/42] Fix CLA_PORCOND_C documentation: C is REAL vector, not DOUBLE PRECISION

+++ lapack/trunk/SRC/cla_porcond_c.f
@@ -38,7 +38,7 @@
 *> \verbatim
 *>
 *>    CLA_PORCOND_C Computes the infinity norm condition number of
-*>    op(A) * inv(diag(C)) where C is a DOUBLE PRECISION vector
+*>    op(A) * inv(diag(C)) where C is a REAL vector
 *> \endverbatim
 *
 *  Arguments:
